Moved logic to use included resources to the Tools project.
