+++ src/Qir/Execution/Tools/Executable/QirFullStateExecutable.cs
@@ -1,8 +1,10 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Reflection;
 using Microsoft.Quantum.Qir.Tools.Driver;
 using Microsoft.Quantum.Qir.Utility;
 
@@ -13,13 +15,7 @@ namespace Microsoft.Quantum.Qir.Tools.Executable
     /// </summary>
     public class QirFullStateExecutable : QirExecutable
     {
-        public QirFullStateExecutable(FileInfo executableFile, byte[] qirBytecode, ILogger logger = null)
-            : base(executableFile,
-                  qirBytecode,
-                  new QirFullStateDriverGenerator(),
-                  logger)
-        {
-        }
+        public override string DriverFileExtension => "cpp";
 
         public override IList<string> LinkLibraries => new List<string> {
                 "Microsoft.Quantum.Qir.Runtime",
@@ -27,6 +23,33 @@ namespace Microsoft.Quantum.Qir.Tools.Executable
                 "Microsoft.Quantum.Qir.QSharp.Core"
             };
 
-        public override string DriverFileExtension => "cpp";
+        public override IList<DirectoryInfo> HeaderDirectories { get; } = new List<DirectoryInfo>();
+
+        public override IList<DirectoryInfo> LibraryDirectories { get; } = new List<DirectoryInfo>();
+
+        public QirFullStateExecutable(FileInfo executableFile, byte[] qirBytecode, ILogger logger = null)
+            : base(executableFile,
+                  qirBytecode,
+                  new QirFullStateDriverGenerator(),
+                  logger)
+        {
+            var thisModulePath = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
+            if (string.IsNullOrWhiteSpace(thisModulePath))
+            {
+                throw new InvalidOperationException("Could not get a path for the current assembly location.");
+            }
+            HeaderDirectories.Add(new DirectoryInfo(Path.Combine(thisModulePath, "runtimes", "any", "native", "include")));
+            HeaderDirectories.Add(new DirectoryInfo(Path.Combine(thisModulePath, "Externals", "CLI11")));
+
+            var osID = Environment.OSVersion.Platform switch
+            {
+                PlatformID.Win32NT => "win-x64",
+                PlatformID.Unix => "linux-x64",
+                PlatformID.MacOSX => "osx-x64",
+                _ => throw new ArgumentException("Unsupported operating system architecture."),
+            };
+            LibraryDirectories.Add(new DirectoryInfo(Path.Combine(thisModulePath, "runtimes", osID, "native")));
+            LibraryDirectories.Add(new DirectoryInfo(Path.Combine(thisModulePath, "Simulators", osID)));
+        }
     }
 }
